Modularized ChatRoom into its own view
